LayoutTests/imported/w3c/web-platform-tests/push-api/idlharness.https.any.serviceworker-expected.txt


PASS idl_test setup
PASS idl_test validation
PASS Partial interface ServiceWorkerRegistration: original interface defined
PASS Partial interface ServiceWorkerRegistration: member names are unique
PASS Partial interface ServiceWorkerGlobalScope: original interface defined
PASS Partial interface ServiceWorkerGlobalScope: valid exposure set
PASS Partial interface ServiceWorkerGlobalScope: member names are unique
PASS WorkerGlobalScope includes WindowOrWorkerGlobalScope: member names are unique
FAIL PushManager interface: existence and properties of interface object assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface object length assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface object name assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: attribute supportedContentEncodings assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: operation subscribe(optional PushSubscriptionOptionsInit) assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: operation getSubscription() assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager interface: operation permissionState(optional PushSubscriptionOptionsInit) assert_own_property: self does not have own property "PushManager" expected property "PushManager" missing
FAIL PushManager must be primary interface of registration.pushManager assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL Stringification of registration.pushManager assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "supportedContentEncodings" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "subscribe(optional PushSubscriptionOptionsInit)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: calling subscribe(optional PushSubscriptionOptionsInit) on registration.pushManager with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "getSubscription()" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: registration.pushManager must inherit property "permissionState(optional PushSubscriptionOptionsInit)" with the proper type assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushManager interface: calling permissionState(optional PushSubscriptionOptionsInit) on registration.pushManager with too few arguments must throw TypeError assert_equals: wrong typeof object expected "object" but got "undefined"
FAIL PushSubscriptionOptions interface: existence and properties of interface object assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface object length assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface object name assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface: attribute userVisibleOnly assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscriptionOptions interface: attribute applicationServerKey assert_own_property: self does not have own property "PushSubscriptionOptions" expected property "PushSubscriptionOptions" missing
FAIL PushSubscription interface: existence and properties of interface object assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface object length assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface object name assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: attribute endpoint assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: attribute expirationTime assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: attribute options assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: operation getKey(PushEncryptionKeyName) assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: operation unsubscribe() assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
FAIL PushSubscription interface: operation toJSON() assert_own_property: self does not have own property "PushSubscription" expected property "PushSubscription" missing
PASS PushMessageData interface: existence and properties of interface object
PASS PushMessageData interface object length
PASS PushMessageData interface object name
PASS PushMessageData interface: existence and properties of interface prototype object
PASS PushMessageData interface: existence and properties of interface prototype object's "constructor" property
PASS PushMessageData interface: existence and properties of interface prototype object's @@unscopables property
PASS PushMessageData interface: operation arrayBuffer()
PASS PushMessageData interface: operation blob()
PASS PushMessageData interface: operation json()
PASS PushMessageData interface: operation text()
PASS PushEvent interface: existence and properties of interface object
PASS PushEvent interface object length
PASS PushEvent interface object name
PASS PushEvent interface: existence and properties of interface prototype object
PASS PushEvent interface: existence and properties of interface prototype object's "constructor" property
PASS PushEvent interface: existence and properties of interface prototype object's @@unscopables property
PASS PushEvent interface: attribute data
PASS PushEvent must be primary interface of new PushEvent("type")
PASS Stringification of new PushEvent("type")
PASS PushEvent interface: new PushEvent("type") must inherit property "data" with the proper type
FAIL PushSubscriptionChangeEvent interface: existence and properties of interface object assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface object length assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface object name assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface: existence and properties of interface prototype object assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface: attribute newSubscription assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent interface: attribute oldSubscription assert_own_property: self does not have own property "PushSubscriptionChangeEvent" expected property "PushSubscriptionChangeEvent" missing
FAIL PushSubscriptionChangeEvent must be primary interface of new PushSubscriptionChangeEvent("pushsubscriptionchange") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PushSubscriptionChangeEvent"
FAIL Stringification of new PushSubscriptionChangeEvent("pushsubscriptionchange") assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PushSubscriptionChangeEvent"
FAIL PushSubscriptionChangeEvent interface: new PushSubscriptionChangeEvent("pushsubscriptionchange") must inherit property "newSubscription" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PushSubscriptionChangeEvent"
FAIL PushSubscriptionChangeEvent interface: new PushSubscriptionChangeEvent("pushsubscriptionchange") must inherit property "oldSubscription" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: Can't find variable: PushSubscriptionChangeEvent"
FAIL ServiceWorkerRegistration interface: attribute pushManager assert_true: The prototype object must have a property "pushManager" expected true got false
FAIL ServiceWorkerRegistration interface: registration must inherit property "pushManager" with the proper type assert_inherits: property "pushManager" not found in prototype chain
PASS ServiceWorkerGlobalScope interface: attribute onpush
FAIL ServiceWorkerGlobalScope interface: attribute onpushsubscriptionchange assert_own_property: The global object must have a property "onpushsubscriptionchange" expected property "onpushsubscriptionchange" missing
PASS ServiceWorkerGlobalScope interface: self must inherit property "onpush" with the proper type
FAIL ServiceWorkerGlobalScope interface: self must inherit property "onpushsubscriptionchange" with the proper type assert_own_property: expected property "onpushsubscriptionchange" missing
